AI Sheets Bot/requirements.txt
requests
python-dotenv
gspread
oauth2client